requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
rocketchat-API==0.6.27
urllib3==1.24.1
websocket_client==0.56.0
flask==1.0.2
